split rp_data_tree into multiple files
